[refactor] customThemeProvider 분리한 내역 원복하여 Providers에 포함
